Separate track initialization launchers from kernels
